+++ utils.h
@@ -3,5 +3,5 @@
 #include <opencv2/opencv.hpp>
 #include <vector>
 
-std::vector<double> parameter_img(cv::Mat img);
-std::vector<double> parameter_img(std::string path);
+std::vector<float> parameter_img(cv::Mat img);
+std::vector<float> parameter_img(std::string path);
